+++ blogFrontend/src/admin/components/layout/AdminHeader.module.scss
@@ -1,0 +1,106 @@
+.header {
+  height: 72px;
+  min-height: 72px;
+  background: rgba(255, 255, 255, 0.95);
+  padding: 0 32px;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  box-shadow: 0 2px 12px rgba(180, 120, 255, 0.08);
+  position: sticky;
+  top: 0;
+  z-index: 5;
+  border-bottom-left-radius: 24px;
+  border-bottom-right-radius: 24px;
+  width: 100%;
+  box-sizing: border-box;
+  max-width: 100%;
+}
+
+.left {
+  display: flex;
+  align-items: center;
+}
+
+.toggleBtn {
+  background: none;
+  border: none;
+  font-size: 18px;
+  cursor: pointer;
+  padding: 10px;
+  color: #a259ff;
+  border-radius: 8px;
+  transition: all 0.3s ease;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.toggleBtn:hover {
+  background: #fbeaff;
+  color: #6c3ec1;
+}
+
+.right {
+  display: flex;
+  align-items: center;
+  gap: 20px;
+}
+
+.userInfo {
+  display: flex;
+  align-items: center;
+  gap: 10px;
+  color: #a259ff;
+  font-size: 15px;
+  font-weight: 600;
+}
+
+.avatar {
+  width: 32px;
+  height: 32px;
+  border-radius: 50%;
+  object-fit: cover;
+}
+
+.username {
+  color: #a259ff;
+  font-size: 15px;
+  font-weight: 600;
+}
+
+.logoutBtn {
+  display: flex;
+  align-items: center;
+  gap: 10px;
+  padding: 8px 22px;
+  background: linear-gradient(90deg, #ffb6b9 0%, #fbeaff 100%);
+  color: #fff;
+  border: none;
+  border-radius: 12px;
+  cursor: pointer;
+  font-size: 15px;
+  font-weight: 600;
+  transition: all 0.3s ease;
+  box-shadow: 0 2px 8px #fbeaff44;
+}
+
+.logoutBtn:hover {
+  background: linear-gradient(90deg, #fbeaff 0%, #ffb6b9 100%);
+  transform: translateY(-2px) scale(1.05);
+  color: #a259ff;
+}
+
+.logoutBtn span {
+  white-space: nowrap;
+}
+
+@media (max-width: 768px) {
+  .header {
+    padding: 0 16px;
+  }
+
+  .username {
+    display: none;
+  }
+}
